+++ kernel/include/util.h
@@ -2,20 +2,22 @@
 
 #pragma once
 
+extern inline void idle();
+
 int strlen(char *str);
 void memcpy(void *src, void *dest, unsigned int amount);
 void backmemcpy(void *srcend, void *destend, unsigned int amount);
 void strcpy(char *src, char *dest);
 unsigned char memcmp(void *src, void *dest, unsigned int amount);
 
-static inline unsigned char bittest(void *src, unsigned int bit);
-static inline unsigned char bitinv(void *src, unsigned int bit);
-static inline unsigned char bitclear(void *src, unsigned int bit);
-static inline unsigned char bitset(void *src, unsigned int bit);
+extern inline unsigned char bittest(void *src, unsigned int bit);
+extern inline unsigned char bitinv(void *src, unsigned int bit);
+extern inline unsigned char bitclear(void *src, unsigned int bit);
+extern inline unsigned char bitset(void *src, unsigned int bit);
 
 unsigned char strcmp(char *src, char *dest);
-static inline void memset(void *dest, unsigned int amount, unsigned char val);
-static inline void memzero(void *dest, unsigned int amount);
+extern inline void memset(void *dest, unsigned int amount, unsigned char val);
+extern inline void memzero(void *dest, unsigned int amount);
 void memrev(char *src, int len, char *dest);
 
 char nybble_to_hex(int num);
@@ -24,7 +26,7 @@ void to_filled_dec(int input, char *out, unsigned char size, char fill);
 void to_dec(int input, char *out);
 unsigned int to_uint(char *input);
 
-static inline char *strcat(char *out, char *in);
+extern inline char *strcat(char *out, char *in);
 
 #define NOT_WAITING 0x00
 #define HAS_INTEGER 0x10
